日常更新

+++ src/main/java/com/anxu/livi/model/vo/goods/GoodsBriefVO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 /**
  * 【商品首页简略信息】
  *
@@ -16,8 +18,8 @@ import lombok.NoArgsConstructor;
 public class GoodsBriefVO {
     private Integer goodsId;//商品id,主键,自增
     private String goodsName;//商品名称
-    private Double goodsPrice;//商品价格
-    private Double goodsOriginalPrice;//商品原价
+    private BigDecimal goodsPrice;//商品价格
+    private BigDecimal goodsOriginalPrice;//商品原价
     private Integer goodsSales;//商品销售量
     private String goodsThumbnail;//商品缩略图
 }
